Revert tdf#97983 - Added localization for numeric types
Instead, do what the bug needs explicitly only for the conversion
functions. Also simplify and extend the related unit tests, and make
it hopefully more obvious what the unit tests are testing.

This reverts commit 4fdc90c51e6a1bbb83c1f1826ad5b90dc1ff0ad6

+++ basic/qa/basic_coverage/da-DK/csng-2.bas
@@ -9,10 +9,11 @@
 Option Explicit
 
 Function doUnitTest() as String
-    Dim A As Double
-    ' In the da-DK locale ',' is the decimal separator and the CDbl
-    ' function is expected to use the locale.
-    A = CDbl("222,222")
+    Dim A As Single
+    ' In the da-DK locale the comma is the decimal separator and the
+    ' period is the grouping separator. The CSng function is
+    ' expected to use the locale. Verify that.
+    A = CSng("222,222")
     If A <> 222.222 Then
         doUnitTest = "FAIL"
     Else
